DemoAjax/python/requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
faiss-cpu
pypdf
python-dotenv
tiktoken
